+++ AmazonChimeSDK/AmazonChimeSDK/audiovideo/DefaultAudioVideoController.swift
@@ -18,6 +18,7 @@ import Foundation
     private let clientMetricsCollector: ClientMetricsCollector
     private var videoClientController: VideoClientController
     private let videoTileController: VideoTileController
+    private let appLifecycleObserver: AppLifecycleObserver
     private var primaryMeetingPromotionObserver: PrimaryMeetingPromotionObserver?
 
     public init(audioClientController: AudioClientController,
@@ -25,6 +26,7 @@ import Foundation
                 clientMetricsCollector: ClientMetricsCollector,
                 videoClientController: VideoClientController,
                 videoTileController: VideoTileController,
+                appLifecycleObserver: AppLifecycleObserver,
                 configuration: MeetingSessionConfiguration,
                 logger: Logger) {
         self.audioClientController = audioClientController
@@ -32,6 +34,7 @@ import Foundation
         self.clientMetricsCollector = clientMetricsCollector
         self.videoClientController = videoClientController
         self.videoTileController = videoTileController
+        self.appLifecycleObserver = appLifecycleObserver
         self.configuration = configuration
         self.logger = logger
     }
@@ -47,6 +50,7 @@ import Foundation
     }
 
     public func start(audioVideoConfiguration: AudioVideoConfiguration) throws {
+        self.appLifecycleObserver.startObserve()
         try audioClientController.start(audioFallbackUrl: configuration.urls.audioFallbackUrl,
                                         audioHostUrl: configuration.urls.audioHostUrl,
                                         meetingId: configuration.meetingId,
@@ -64,6 +68,7 @@ import Foundation
     public func stop() {
         audioClientController.stop()
         videoClientController.stopAndDestroy()
+        self.appLifecycleObserver.stopObserve()
     }
 
     public func addAudioVideoObserver(observer: AudioVideoObserver) {