+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## PENDING
 
 ### Added
-* Added meetingReconnected/audioInputFailed/signalingDropped meeting event
+* Added meetingReconnected/audioInputFailed/signalingDropped/appEnteredForeground/appEnteredBackground meeting event
 * Added meetingReconnectDurationMs/audioInputErrorMessage/signalingDroppedErrorMessage meeting event attributes
 
 ## [0.27.1] - 2025-03-28